refactor: move topic utils to custom bouquet
