+++ src/custom/ecospheres/views/bouquets/BouquetDetailView.vue
@@ -15,6 +15,10 @@ import OrganizationLogo from '@/components/OrganizationLogo.vue'
 import ReusesList from '@/components/ReusesList.vue'
 import config from '@/config'
 import BouquetDatasetList from '@/custom/ecospheres/components/BouquetDatasetList.vue'
+import {
+  useClonedFrom,
+  useBreadcrumbLinks
+} from '@/custom/ecospheres/utils/bouquet'
 import type { Topic, DatasetProperties } from '@/model/topic'
 import { useTopicStore } from '@/store/TopicStore'
 import { useUserStore } from '@/store/UserStore'
@@ -22,7 +26,6 @@ import { descriptionFromMarkdown, formatDate } from '@/utils'
 import { getOwnerAvatar } from '@/utils/avatar'
 import { useSpatialCoverage } from '@/utils/spatial'
 import { getThemeTextColor, getThemeColor } from '@/utils/theme'
-import { useClonedFrom, useBreadcrumbLinks } from '@/utils/topic'
 
 import BouquetDatasetListExport from '../../components/BouquetDatasetListExport.vue'
